requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.7.0
Flask-WTF==1.2.1
python-dotenv==1.0.1
Flask-Migrate==4.0.5
